Add error checking to api-request
Looks to see if the response from the server comes back http 200 before
trying to parse the body as json. If any other response code is returned
with throw an exception.

Also, adds docuementation to the primary entry point
'spoon.core/api-request, and adds a helper for creating the minimal
required options with the 'spoon.core/client-info function.